+++ xscen/biasadjustment.py
@@ -7,8 +7,8 @@ from xclim import sdba
 from xclim.core.calendar import convert_calendar, get_calendar
 
 from .catalog import parse_from_ds
-from .common import minimum_calendar
 from .config import parse_config
+from .utils import minimum_calendar
 
 # TODO: Change all paths to PosixPath objects, including in the catalog?
 # TODO: Compute sometimes fails randomly (in debug, pretty much always). Also (detrend?) fails with pr. Investigate why.
